pyproject.toml
[project]
name = "multimodal-rag-chatbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "vosk>=0.3.45",
    "chromadb>=0.4.15",
    "pymupdf>=1.23.0",
    "google-generativeai>=0.3.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "moviepy==1.0.3",
    "numpy>=1.24.0",
    "pandas>=2.1.0",
    "aiofiles>=23.2.0",
    "soundfile>=0.12.1",
    "scipy>=1.11.0",
    "sentence-transformers>=2.2.0",
]
